Finalizing display of canvas-rendered scatterplots.
Canvas-rendered scatterplots need to be called from a containing div to keep canvas display of elements and interaction through SVG aligned on the webpage.  No interaction is implemented yet, and the jury is still out for how to redraw the canvas points (especially since `renderPoints` is an internal method within `redrawCanvas()`).